♻️ Refactor: my-review의 탭 동작 수정
