+++ src/app/(crew)/my-page/my-review/page.tsx
@@ -1,0 +1,3 @@
+export default function MyReviewPage() {
+  return <div>my review list component</div>;
+}
