+++ src/app/(crew)/my-page/reviewable/page.tsx
@@ -1,0 +1,3 @@
+export default function ReviewableGatheringsPage() {
+  return <div>reviewable gathering list component</div>;
+}
